[NoSQL] Make TransactionalPersistence an interface
This change makes `TransactionalPersistence` an interface and introduces `AbstractTransactionalPersistence` having the code that was in `TransactionalPersistence`.
